requirements.txt
atlassian-python-api==3.19.0
certifi==2021.10.8
charset-normalizer==2.0.11
Deprecated==1.2.13
idna==3.3
oauthlib==3.2.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
urllib3==1.26.8
wrapt==1.13.3
